Removed unneeded folders and renamed the multiagent suppression file.
