--- inputs/metadata.in
@@ -1,23 +1,0 @@
-NAME = tutorial03
-FIRENAME = tutorial03
-CENTER_LAT = 38.279998779296875
-CENTER_LON = -120.08000183105469
-WEST_BUFFER = 30000.0
-EAST_BUFFER = 30000.0
-SOUTH_BUFFER = 30000.0
-NORTH_BUFFER = 30000.0
-DO_FUEL = True
-FUEL_SOURCE = landfire
-FUEL_VERSION = 2.3.0
-DO_WX = False
-WX_TYPE = historical
-WX_START_TIME = 2011-01-30 00:00
-WX_NUM_HOURS = 0
-DO_IGNITION = False
-POINT_IGNITION = True
-POLYGON_IGNITION = False
-IGNITION_LAT = -9999.0
-IGNITION_LON = -9999.0
-ACTIVE_FIRE_TIMESTAMP = null
-ALREADY_BURNED_TIMESTAMP = null
-IGNITION_RADIUS = -9999.0
