--- inputs/metadata.out
@@ -1,7 +1,0 @@
-SRS = EPSG:32610
-DOMAIN_TE = 725416.164839092 4210916.19800615 785416.164839092 4270916.19800615
-WX_TE = 719416.164839092 4204916.19800615 791416.164839092 4276916.19800615
-WX_BEGIN =  2011-01-30 00:00
-WX_END = 2011-01-30 00:00
-CENTER_LON = -120.08000183105469
-CENTER_LAT = 38.279998779296875
