+++ multisuppressant.py
